Added more flexibility for fragment use.
In particular, now you can use fragments in the middle of
a line, and use more than one fragment in a line.

I also decided to make indentation automatic on the perl
fragments.  So, $indentation is no longer available to the
user-provided code.  I don't think that will ever be a problem,
as you can just put un-indended fragment uses in the source and
get the same effect.  I think I was being too fancy.